refactor(`ft_printf`): Use `uintptr_t` for `%p`
`uintptr_t` is capable of holding any pointer type, but it is an optional type. A guaranteed alternative would be `uintmax_t`. But for now, I like that `uintptr_t` communicates that it stores a pointer address. If it is not available, pointers on that system might have some metadata and could not be easily printed anyway.